AppStore: split into 5 domain extensions
AppStore.swift was 691 lines, well past the 600-line hard cap, holding
the @Observable class plus event handling, notification queue management,
OTA orchestration, log buffering, and device lookups all in one file.

Splits methods into domain-scoped extensions while keeping stored
properties (which @Observable requires in the main class) and the
init/reset lifecycle in the main file:

- AppStore.swift (116) — class declaration, all stored properties, the
  DeviceCheckResult enum, init, reset, and the first-seen persistence
  helpers (recordFirstSeen / removeFirstSeen are now internal so the
  Events extension can call them).
- AppStore+Events.swift (275) — the apply(_ event:) dispatcher plus the
  static logEntry(from:) and notification(from:) helpers it uses.
- AppStore+Notifications.swift (109) — pop / popFastTrack /
  enqueueOTABulkSummary / enqueueNotification / notification(for:) /
  stripped helper.
- AppStore+OTA.swift (140) — activeOTAUpdates, otaStatus(for:), the
  start/cancel methods, and the handle… response/state callbacks plus
  flashCheckResult.
- AppStore+Logs.swift (22) — clearLogs, insertLogEntry, insertRawLogEntry.
- AppStore+Devices.swift (49) — device(named:), state(for:), isAvailable,
  and the optimisticRename / revertOptimisticRename pair.

A handful of formerly-private members (pendingRenames, recordFirstSeen,
removeFirstSeen, revertOptimisticRename, the OTA handlers, etc.) bumped
to internal so cross-file extensions can reach them. No external API
surface change — all of these were already module-internal, just
file-private within AppStore.swift.

The five new files are quoted in pbxproj (the `+` requires it) and added
to the ShellbeeWidgetsExtension membership exception list.

Build green. No behavior changes intended.

Refs #34
